Issue: #7 Added Watchdog for our IRC connection.
Every minute we send a ping.  We wait for a minute for a PONG from the server.  If, after a minute, we don't hear from the server, we reconnect after a minute, and keep trying.

Our WatchDog is now the driver behind reconnects, it no longer happens on the reader thread.