Issue: #5 Added Cli Installer.
This takes four arguments:
 - Solution Directory
 - The root where to put the Chaskis folder for install
 - The path to the WIX XML folder.
 - Which target (Debug or Release) we are doing.

What this program does is parse the WIX folder and install everything in the specified root directory.  This will be useful during PKGBUILD so we don't need to copy everything manually, just run this program with the correct arguments.  It also means we only need one XML File for installing.

Let's hope this crazy thing works...